move logo to UI folder
