+++ src/components/ui/Logo.tsx
@@ -3,11 +3,11 @@ import { Link } from 'react-router-dom';
 type LogoSize = 'xs' | 'sm' | 'md' | 'lg' | 'xl';
 
 const sizeClasses: Record<LogoSize, string> = {
-  xs: 'h-2',   // 32px
-  sm: 'h-4',  // 40px
-  md: 'h-6',  // 48px
-  lg: 'h-8',  // 56px
-  xl: 'h-10',  // 64px
+  xs: 'h-4',   // 16px
+  sm: 'h-5',   // 20px
+  md: 'h-6',   // 24px
+  lg: 'h-7',   // 28px
+  xl: 'h-8',   // 32px
 };
 
 export function Logo({ size = 'md', className = '' }: { size?: LogoSize; className?: string }) {
